requirements.txt
fastapi
uvicorn[standard]
pydantic
dotenv
google-genai